pyproject.toml
[tool.poetry]
name = "wealthbox"
version = "0.8.0"
description = ""
authors = ["spencerogden-dsam <67068943+spencerogden-dsam@users.noreply.github.com>"]

[tool.semantic_release]
version_variable = "pyproject.toml:version"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
click = {version = "^8.0", optional = true}
rich = {version = "^13.0", optional = true}
python-dotenv = {version = "^1.0", optional = true}

[tool.poetry.extras]
cli = ["click", "rich", "python-dotenv"]

[tool.poetry.scripts]
wb = "wealthbox.cli.main:cli"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
responses = "^0.25.0"

[tool.poetry.group.dev.dependencies]
python-semantic-release = "^9.4.1"

[tool.poetry.group.cli.dependencies]
click = "^8.0"
rich = "^13.0"
python-dotenv = "^1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
